requirements.txt
streamlit
langchain
python-dotenv
langchain-groq
replicate
fpdf
python-docx
scikit-learn
nltk
